Change to work with new data file

+++ src/main/java/com/example/springboot/Services/MapServiceBuilder.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import com.example.springboot.Models.MapNode;
 
 public interface MapServiceBuilder {
-  public Map<Integer, MapNode> getNodeMap();
-  public Map<Integer, List<Integer>> getEdgeMap();
+  public Map<Long, MapNode> getNodeMap();
+  public Map<Long, List<Long>> getEdgeMap();
   public void buildMap();
 }
